refactor: changed architecture of directories
