+++ moutarde-back/moutarde-back/Infrastructure/Data/MoutardeDbContext.cs
@@ -1,6 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 
-namespace moutarde_back.Data;
+namespace moutarde_back.Infrastructure.Data;
 
 public class MoutardeDbContext(DbContextOptions<MoutardeDbContext> options) : DbContext(options)
 {
