+++ moutarde-back/moutarde-back/moutarde-back.csproj
@@ -18,4 +18,8 @@
         <PackageReference Include="Swashbuckle.AspNetCore" Version="10.1.2" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Infrastructure\" />
+    </ItemGroup>
+
 </Project>
